Added simple menu items add function

+++ venues/views.py
@@ -12,12 +12,14 @@ from .models import (Venue,
                      Comment,
                      Rating,
                      Review,
-                     Menu,)
+                     Menu,
+                     MenuItems,)
 from .forms import (VenueForm,
                     CommentForm,
                     ReviewForm,
                     RatingForm,
-                    MenuForm,)
+                    MenuForm,
+                    MenuItemsFormSet,)
 
 
 def create_session(request):
@@ -129,19 +131,19 @@ def create_venue(request):
     purpose = "Dodaj"
     helper_text = "Po stworzeniu swojego lokalu, będzie on widoczny tylko dla Ciebie. Można go aktywować w opcji Edytuj"
 
-    form = VenueForm()
-    is_active_field = form.fields['is_active']
+    venue_form = VenueForm()
+    is_active_field = venue_form.fields['is_active']
     is_active_field.disabled = True
     is_active_field.widget = is_active_field.hidden_widget()
     if request.method == 'POST':
-        form = VenueForm(request.POST)
-        if form.is_valid():
-            venue = form.save(commit=False)
+        venue_form = VenueForm(request.POST)
+        if venue_form.is_valid():
+            venue = venue_form.save(commit=False)
             venue.owner = request.user
             venue.save()
-            venue.categories.add(*form.cleaned_data['categories'])
+            venue.categories.add(*venue_form.cleaned_data['categories'])
             return redirect('venues:venues')
-    context = {'form': form, 'helper_text': helper_text, 'purpose':purpose}
+    context = {'venue_form': venue_form, 'helper_text': helper_text, 'purpose':purpose}
     return render(request, 'venues/venue_form.html', context)
 
 
@@ -217,6 +219,7 @@ def create_review(request, pk):
     context = {'review_form': review_form, 'rating_form': rating_form}
     return render(request, 'venues/review_form.html', context)
 
+@login_required(login_url='venues:login')
 def profile(request, pk):
     user = User.objects.get(id = pk)
     venues = Venue.objects.filter(owner = user.id).order_by('-is_active','name')
@@ -225,12 +228,11 @@ def profile(request, pk):
     context = {'venues': venues, 'reviews': reviews, 'avg_rating':avg_rating}
     return render(request, 'venues/profile.html', context)
 
-
+@login_required(login_url='venues:login')
 def menu_list(request, pk):
     venue = Venue.objects.get(id=pk)
     menus = venue.menus.all()
     menu_form = MenuForm()
-    
     if request.method == 'POST' and 'menu_add' in request.POST:
         Menu.objects.create(
             name = request.POST.get('name'),
@@ -238,17 +240,18 @@ def menu_list(request, pk):
             venue=venue
             )
         return redirect('venues:menu', pk=venue.pk)
-
         
-        
-    context = {'venue': venue, 'menus': menus, 'menu_form': menu_form,}
+    context = {'venue': venue, 'menus': menus, 'menu_form': menu_form}
     return render(request, 'venues/menu.html', context)
 
-def delete_menu(request, menu):
-    pass
-
+@login_required(login_url='venues:login')
 def delete_menu(request, pk):
     menu = Menu.objects.get(id=pk)
+    
+    if request.user != menu.venue.owner:
+        messages.error(request, "You do not have permission to delete this comment")
+        return redirect('venues:menu', pk = menu.venue.pk)
+    
     if request.method == 'POST':
         menu.delete()
         return redirect('venues:menu', pk = menu.venue.pk)
@@ -256,3 +259,15 @@ def delete_menu(request, pk):
         return render(request, 'venues/detail.html') 
 
 
+def menu_items(request, pk, menu):
+    venue = Venue.objects.get(id=pk)
+    menu = Menu.objects.get(id=menu)
+    formset = MenuItemsFormSet(instance = menu)
+    if request.method == 'POST' and 'menu_add_items' in request.POST:
+        formset = MenuItemsFormSet(request.POST, instance=menu)
+        if formset.is_valid():
+            formset.save()
+            return redirect('venues:menu', pk=venue.pk)
+
+    context = {'menu': menu, 'formset': formset,}
+    return render(request, 'venues/menu_items_form.html', context)
